pyproject.toml
[tool.poetry]
name = "scurri"
version = "0.1.0"
description = ""
authors = ["Luke Shiner <luke@lukeshiner.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
python-dateutil = "^2.8.2"
types-python-dateutil = "^0.1.4"

[tool.poetry.dev-dependencies]
isort = "^5.9.3"
black = "^21.7b0"
flake8 = "^3.9.2"
pytest = "^6.2.4"
mypy = "^0.910"
coveralls = "^3.2.0"
requests-mock = "^1.9.3"
pytest-cov = "^2.12.1"
flake8-docstrings = "^1.6.0"
types-requests = "^2.25.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
